requirements.txt
gradio
faiss-cpu
pandas
sentence-transformers
